+++ app/orgs/orgs-data.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import { useSearchParams, useRouter } from 'next/navigation'
 import { useState, useEffect } from 'react'
 import Link from '@components/common/link'
-import { CATEGORIES } from './page'
+import { CATEGORIES } from './categories'
 import styles from './page.module.scss'
 import useClient from 'hooks/client'
 import { useSmallerThan } from 'hooks/useWindowSize'
